Use mruby-require to load files.
This is preferred over the old "grab the entire folder, meh" way,
because the old one had no way of tracking dependencies and loaded
files alphabetically, meaning if one file was required before the
other, all hell would break loose. The gem has small glitches
that were added as a TODO, but for current needs it "just works".