Texture rendering on crack.
Updating a VBO on the fly N times for N sprites per frame is not
efficient, especially since it's the same most of the time. This
update temporarily removes clipping, because clipping is basically
the only time where we need to update the VBO (apart from the
initial construction of the VBO). So I propose clipping calls a
method that uses the specified clip to update the VBO just once.

Note that the position of the rendered sprite is not determined
by the vertices, those are there just for the width and height
dimensions, the actual positioning of the sprite is done via a
translation matrix (model matrix, converting from model to world
coordinates) in the shader.