Move VBO/IBO from Texture to Sprite.
First in the set of changes. A texture has no business rendering,
it just holds a single texture. Sprite and Spritesheet (which
should be a subclass of Sprite) should have it's own VBO/IBO
handling and rendering. Possibly introduce a Renderer object
with draw_VBO(), but then per Shader variables will need to be
set up by magic somehow.