Progress...
Android is an annoying fuck. We have to wait for a window to be
constructed, only then can we start opengl and mrb and rendering.

Even more annoying, any time the app goes out of focus, our opengles
context may be destroyed and all of the textures, shaders and vbos
alongside it...