Added linked lists
This is done using a LinkNode.
A LinkNode is a wrapper around any object, and allows the linkage of other LinkNodes to formed a linked list.

LinkNodes are enumerable, and can be looped.
```ruby
LinkNode#loop!
```
Beware, most enumerable methods will continue to loop if enumerating a looped linked list.

The fastest method for creating a linked list is to use Array#to_linked_list

```ruby
[1, 2, 3].to_linked_list
```